+++ H6Game/H6Game.ServerApp/TestBenckmark.cs
@@ -48,7 +48,7 @@ namespace H6Game.ServerApp
 
         public static async Task Call(Network network)
         {
-            await network.CallMessageAsync<TestMessage, TestMessage>(send, TestCMD.TestCmd);
+            await network.CallMessageAsync<TestMessage, TestMessage>(send, (ushort)TestCMD.TestCmd);
 
             Count++;
             if (stopWatch.ElapsedMilliseconds > 1000)
