+++ H6Game/H6Game.Test/UnitTest/H6Game.TesterSubscriber/NetSubscribers.cs
@@ -12,7 +12,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubEmpty)]
     public class SubEmpty : NetSubscriber
     {
-        protected override void Subscribe(Network network, int netCommand)
+        protected override void Subscribe(Network network, ushort netCommand)
         {
             network.Response(1024);
         }
@@ -21,7 +21,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubInt)]
     public class SubInt : NetSubscriber<int>
     {
-        protected override void Subscribe(Network network, int message, int netCommand)
+        protected override void Subscribe(Network network, int message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -30,7 +30,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubUInt)]
     public class SubUInt : NetSubscriber<uint>
     {
-        protected override void Subscribe(Network network, uint message, int netCommand)
+        protected override void Subscribe(Network network, uint message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -39,7 +39,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubString)]
     public class SubString : NetSubscriber<string>
     {
-        protected override void Subscribe(Network network, string message, int netCommand)
+        protected override void Subscribe(Network network, string message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -48,7 +48,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubLong)]
     public class SubLong : NetSubscriber<long>
     {
-        protected override void Subscribe(Network network, long message, int netCommand)
+        protected override void Subscribe(Network network, long message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -57,7 +57,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubULong)]
     public class SubULong : NetSubscriber<ulong>
     {
-        protected override void Subscribe(Network network, ulong message, int netCommand)
+        protected override void Subscribe(Network network, ulong message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -66,7 +66,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubFloat)]
     public class SubFloat : NetSubscriber<float>
     {
-        protected override void Subscribe(Network network, float message, int netCommand)
+        protected override void Subscribe(Network network, float message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -75,7 +75,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubDouble)]
     public class SubDouble : NetSubscriber<double>
     {
-        protected override void Subscribe(Network network, double message, int netCommand)
+        protected override void Subscribe(Network network, double message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -84,7 +84,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubDecimal)]
     public class SubDecimal : NetSubscriber<decimal>
     {
-        protected override void Subscribe(Network network, decimal message, int netCommand)
+        protected override void Subscribe(Network network, decimal message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -93,7 +93,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubByte)]
     public class SubByte : NetSubscriber<byte>
     {
-        protected override void Subscribe(Network network, byte message, int netCommand)
+        protected override void Subscribe(Network network, byte message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -102,7 +102,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubSByte)]
     public class SubSByte : NetSubscriber<sbyte>
     {
-        protected override void Subscribe(Network network, sbyte message, int netCommand)
+        protected override void Subscribe(Network network, sbyte message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -111,7 +111,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubShort)]
     public class SubShort : NetSubscriber<short>
     {
-        protected override void Subscribe(Network network, short message, int netCommand)
+        protected override void Subscribe(Network network, short message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -120,7 +120,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubUShort)]
     public class SubUShort : NetSubscriber<ushort>
     {
-        protected override void Subscribe(Network network, ushort message, int netCommand)
+        protected override void Subscribe(Network network, ushort message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -129,7 +129,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubChar)]
     public class SubUChar : NetSubscriber<char>
     {
-        protected override void Subscribe(Network network, char message, int netCommand)
+        protected override void Subscribe(Network network, char message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -138,7 +138,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubClass)]
     public class SubObject : NetSubscriber<TesterMessage>
     {
-        protected override void Subscribe(Network network, TesterMessage message, int netCommand)
+        protected override void Subscribe(Network network, TesterMessage message, ushort netCommand)
         {
             network.Response(message);
         }
@@ -147,7 +147,7 @@ namespace H6Game.TesterSubscriber
     [NetCommand(NetCommandTest.SubEnum)]
     public class SubEnum : NetSubscriber<EnumType>
     {
-        protected override void Subscribe(Network network, EnumType message, int netCommand)
+        protected override void Subscribe(Network network, EnumType message, ushort netCommand)
         {
             network.Response(message);
         }
