+++ H6Game/H6Game.Test/UnitTest/H6Game.TesterSubscriber/Program.cs
@@ -51,7 +51,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var emptyResult = await network.CallMessageAsync<int>(NetCommandTest.SubEmpty);
+            var emptyResult = await network.CallMessageAsync<int>((ushort)NetCommandTest.SubEmpty);
             if(1024 == emptyResult)
             {
                 Console.WriteLine("SubEmpty success.");
@@ -65,7 +65,7 @@ namespace H6Game.TesterSubscriber
         private static async void SubInt()
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
-            var intResult = await network.CallMessageAsync<int, int>(1024, NetCommandTest.SubInt);
+            var intResult = await network.CallMessageAsync<int, int>(1024, (ushort)NetCommandTest.SubInt);
             if (1024 == intResult)
             {
                 Console.WriteLine("SubInt success.");
@@ -80,7 +80,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var uintResult = await network.CallMessageAsync<uint, uint>(1024, NetCommandTest.SubUInt);
+            var uintResult = await network.CallMessageAsync<uint, uint>(1024, (ushort)NetCommandTest.SubUInt);
             if (1024 == uintResult)
             {
                 Console.WriteLine("SubUInt success.");
@@ -95,7 +95,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var stringResult = await network.CallMessageAsync<string, string>("1024", NetCommandTest.SubString);
+            var stringResult = await network.CallMessageAsync<string, string>("1024", (ushort)NetCommandTest.SubString);
             if ("1024" == stringResult)
             {
                 Console.WriteLine("SubString success.");
@@ -110,7 +110,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var longResult = await network.CallMessageAsync<long, long>(102410241024, NetCommandTest.SubLong);
+            var longResult = await network.CallMessageAsync<long, long>(102410241024, (ushort)NetCommandTest.SubLong);
             if (102410241024 == longResult)
             {
                 Console.WriteLine("SubLong success.");
@@ -125,7 +125,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var ulongResult = await network.CallMessageAsync<ulong, ulong>(102410241024, NetCommandTest.SubULong);
+            var ulongResult = await network.CallMessageAsync<ulong, ulong>(102410241024, (ushort)NetCommandTest.SubULong);
             if (102410241024 == ulongResult)
             {
                 Console.WriteLine("SubULong success.");
@@ -140,7 +140,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var floatResult = await network.CallMessageAsync<float, float>(1024.1024f, NetCommandTest.SubFloat);
+            var floatResult = await network.CallMessageAsync<float, float>(1024.1024f, (ushort)NetCommandTest.SubFloat);
             if (1024.1024 == floatResult)
             {
                 Console.WriteLine("SubFloat success.");
@@ -155,7 +155,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var doubleResult = await network.CallMessageAsync<double, double>(10241024.1024, NetCommandTest.SubDouble);
+            var doubleResult = await network.CallMessageAsync<double, double>(10241024.1024, (ushort)NetCommandTest.SubDouble);
             if (10241024.1024 == doubleResult)
             {
                 Console.WriteLine("SubDouble success.");
@@ -170,7 +170,7 @@ namespace H6Game.TesterSubscriber
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
-            var decimalResult = await network.CallMessageAsync<decimal, decimal>(10241024.1024m, NetCommandTest.SubDecimal);
+            var decimalResult = await network.CallMessageAsync<decimal, decimal>(10241024.1024m, (ushort)NetCommandTest.SubDecimal);
             if (10241024.1024m == decimalResult)
             {
                 Console.WriteLine("SubDecimal success.");
@@ -186,7 +186,7 @@ namespace H6Game.TesterSubscriber
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
             byte bt = 127;
-            var byteResult = await network.CallMessageAsync<byte,byte>(bt, NetCommandTest.SubByte);
+            var byteResult = await network.CallMessageAsync<byte,byte>(bt, (ushort)NetCommandTest.SubByte);
             if (bt == byteResult)
             {
                 Console.WriteLine("SubByte success.");
@@ -202,7 +202,7 @@ namespace H6Game.TesterSubscriber
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
             sbyte sbt = 127;
-            var sbyteResult = await network.CallMessageAsync<sbyte, sbyte>(sbt, NetCommandTest.SubSByte);
+            var sbyteResult = await network.CallMessageAsync<sbyte, sbyte>(sbt, (ushort)NetCommandTest.SubSByte);
             if (sbt == sbyteResult)
             {
                 Console.WriteLine("SubSByte success.");
@@ -218,7 +218,7 @@ namespace H6Game.TesterSubscriber
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
             short st = 30000;
-            var shortResult = await network.CallMessageAsync<short, short>(st, NetCommandTest.SubShort);
+            var shortResult = await network.CallMessageAsync<short, short>(st, (ushort)NetCommandTest.SubShort);
             if (st == shortResult)
             {
                 Console.WriteLine("SubShort success.");
@@ -234,7 +234,7 @@ namespace H6Game.TesterSubscriber
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
 
             ushort ust = 65535;
-            var ushortResult = await network.CallMessageAsync<ushort, ushort>(ust, NetCommandTest.SubUShort);
+            var ushortResult = await network.CallMessageAsync<ushort, ushort>(ust, (ushort)NetCommandTest.SubUShort);
             if (ust == ushortResult)
             {
                 Console.WriteLine("SubUShort success.");
@@ -244,7 +244,7 @@ namespace H6Game.TesterSubscriber
         private static async void SubChar()
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
-            var charResult = await network.CallMessageAsync<char, char>('a', NetCommandTest.SubSByte);
+            var charResult = await network.CallMessageAsync<char, char>('a', (ushort)NetCommandTest.SubSByte);
             if ('a' == charResult)
             {
                 Console.WriteLine("SubChar success.");
@@ -263,7 +263,7 @@ namespace H6Game.TesterSubscriber
                 Message = "测试消息！",
                 TestId = 1024,
             };
-            var charResult = await network.CallMessageAsync<TesterMessage, TesterMessage>(message, NetCommandTest.SubClass);
+            var charResult = await network.CallMessageAsync<TesterMessage, TesterMessage>(message, (ushort)NetCommandTest.SubClass);
             if (message.Message == charResult.Message && message.TestId == charResult.TestId)
             {
                 Console.WriteLine("SubClass success.");
@@ -277,7 +277,7 @@ namespace H6Game.TesterSubscriber
         private static async void SubEnum()
         {
             var network = Game.Scene.GetComponent<NetConnectorComponent>().Network;
-            var enumResult = await network.CallMessageAsync<EnumType, EnumType>(EnumType.Test1, NetCommandTest.SubEnum);
+            var enumResult = await network.CallMessageAsync<EnumType, EnumType>(EnumType.Test1, (ushort)NetCommandTest.SubEnum);
             if (EnumType.Test1 == enumResult)
             {
                 Console.WriteLine("SubEnum success.");
